Updated admin allshift.show and edit to include all fields

+++ NRSN-SMS/database/migrations/2023_08_07_054728_create_shifts_table.php
@@ -20,9 +20,9 @@ return new class extends Migration
             $table->boolean('isflagged')->default(0);
             $table->boolean('isinvoiced')->default(0);
             $table->date('date');
-            $table->float('expenses');
+            $table->integer('expenses');
             $table->integer('Km');
-            $table->float('hours');
+            $table->integer('hours');
             $table->timestamps();
         });
     }
